+++ CHANGELOG.md
@@ -9,6 +9,7 @@ While backward compatibility is maintained for now, users are encouraged to upda
 
 - `$anchor` support.
 - `$recursiveRef` & `$recursiveAnchor` support in Draft 2019-09.
+- `$dynamicRef` & `$dynamicAnchor` support in Draft 2020-12.
 
 ### Changed